Adicionando um página para produtos e outra para o carrinho
