+++ src/components/ProdutoContainerCarrinho.js
@@ -12,24 +12,25 @@ const Produto = styled.div`
     grid-auto-flow: column;
     gap: 8px;
     align-items: center; 
-`
+`;
 
 class ProdutoContainerCarrinho extends React.Component {
   render() {
     return (
       <ContainerCarrinho>
-        {this.props.carrinho.map((produto) => {
-          return (
-            <Produto>
-              <p>{produto.quantidade}x</p>
-              <p>{produto.nome}</p>
-              {/* <p>R$ {produto.valor}</p> */}
-              <button onClick={() => this.props.deletarProduto(produto.id)}>
-                Remover
-              </button>
-            </Produto>
-          );
-        })}
+        {this.props.carrinho.map((produto) => (
+          <Produto key={produto.key}>
+            <p>
+              {produto.quantidade}
+              x
+            </p>
+            <p>{produto.nome}</p>
+            {/*<p>R$ {produto.valor}</p> */}
+            <button onClick={() => this.props.deletarProduto(produto.id)}>
+              Remover
+            </button>
+          </Produto>
+        ))}
       </ContainerCarrinho>
     );
   }
